ScreenshotSystem

+++ samples/run-all.sh
@@ -5,6 +5,7 @@ echo "[gengine-samples] Running all samples ..."
 for dir in `ls`;
 do
     test -d "$dir" || continue
+    test -e "$dir/build.hxml" || continue
     cd $dir
     echo "[gengine-samples] Running <$dir> ..."
     gengine-run $1
